XMLHttpRequest/open-url-fragment.htm

<!DOCTYPE html>
<html>
  <head>
    <title>XMLHttpRequest: open() resolving URLs - fragment identifier</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(function() {
        var client = new XMLHttpRequest()
        client.open("GET", "folder.txt#foobar", false)
        client.send(null)
        assert_equals(client.responseText, "top\n")
      })
    </script>
  </body>
</html>
